refactor:공유 API mapper 유틸 분리
