+++ src/shared/api/sse/sseStepMapper.ts
